% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentences.R
\name{get_sentences}
\alias{get_sentences}
\title{Function extracting sentences form raw text}
\usage{
get_sentences(
  text,
  language,
  lem = FALSE,
  remove_no = FALSE,
  remove_punct = FALSE,
  tolower = FALSE,
  verbose = FALSE,
  n_cores = 1
)
}
\arguments{
\item{text}{Vector of strings that is going to be tokenized.}

\item{language}{Language model that is used for tokenization. See language models at https://github.com/bnosac/udpipe.}

\item{lem}{Logical parameter for extracting also lemmatized version of a sentence. Default is FALSE.}

\item{remove_no}{Logical parameter for removing numbers. Default is FALSE.}

\item{remove_punct}{Logical parameter for removing punctuation. Default is FALSE.}

\item{tolower}{Logical parameter for transforming strings to lower case. Default is FALSE.}

\item{verbose}{Logical parameter for displaying extended information on processed data. Works only with processing on one core. Default is FALSE.}

\item{n_cores}{Numeric parameter for number of cores to be used for processing. Default is 1 core.}
}
\description{
This function allows tokenizing text on the level of sentences.
}
\examples{
get_sentences()
}
\keyword{agnostic}
\keyword{language}
\keyword{sentence}
\keyword{tokenization}
