% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent_ngrams.R
\name{sent_ngrams}
\alias{sent_ngrams}
\title{Function for generating sentence level n-grams - raw sentences}
\usage{
sent_ngrams(sentences, n)
}
\arguments{
\item{sentences}{Table with raw sentences generated by get_sentences function. No default.}

\item{n}{Integer specifying the parsing rule for ngrams: 2 for bigrams, 3 for trigrams, etc. No default.}
}
\description{
This function allows tokenizing text on the level of sentences.
}
\examples{
sent_ngrams()
}
\keyword{sentences}
\keyword{tokenization}
