% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent_ngrams_lem.R
\name{sent_ngrams_lem}
\alias{sent_ngrams_lem}
\title{Function for generating sentence level n-grams - lemmatized sentences}
\usage{
sent_ngrams_lem(sentences, n)
}
\arguments{
\item{sentences}{Table with lemmatized sentences generated by get_sentences function. No default.}

\item{n}{Integer specifying the parsing rule for ngrams: 2 for bigrams, 3 for trigrams, etc. No default.}
}
\description{
This function allows tokenizing text on the level of sentences.
}
\examples{
sent_ngrams_lem()
}
\keyword{lemmatization}
\keyword{sentences}
\keyword{tokenization}
